Ai_assistant-memory-voice/requirements.txt
google-genai
python-dotenv
SpeechRecognition
gTTS
pyaudio
